Add min_branch_date param for into_branches option
It is common when we only want to process an interval on master branch along with
branches merged back into master during this period. However, if a branch is created
long before the start of the interval on master, the existing code wouldn't work
correctly. Thus, we add a min_branch_date option so that we know when to stop during
backtracing of 2nd phase. In most common use case, min_branch_date should be set to
the start time of the interval on master.