Fix 2 minor bugs in 2nd phase
The first bug is that Processor used to perform checking after processing.
So if the first commit in a new branch is invalid, we would still process that commit.
The second bug is about branch_cnt. Before the fix, branch_cnt would be updated
even if the first commit of a potential branch is invalid (has been visited before, etc).