Add java server and utillity function to support
This commit has java server implementation which is responsible to communicate
with the language agnostic commit parser and carry out the call commit graph creation.

The call-graph files are support files to parse java grammar and create call commit graph
update it and also update the stats for a given file. This is heavily inspired from the c
implementations.

Signed-off-by: Farhaan Bukhsh <farhaan.bukhsh@gmail.com>